rename: distribution files
distributionParameters folder -> distriubtion

distributionMain.py -> ui.py

distributionFunctions.py -> utils.py
