+++ src/python/impactx/dashboard/Input/distribution/__init__.py
@@ -1,0 +1,5 @@
+from .utils import DistributionFunctions
+
+__all__ = [
+    "DistributionFunctions",
+]
