--- src/python/impactx/dashboard/Input/distributionParameters/__init__.py
@@ -1,5 +1,0 @@
-from .distributionFunctions import DistributionFunctions
-
-__all__ = [
-    "DistributionFunctions",
-]
