+++ src/python/impactx/dashboard/Run/simulation.py
@@ -7,7 +7,7 @@ License: BSD-3-Clause-LBNL
 """
 
 from .. import setup_server
-from ..Input.distributionParameters.distributionFunctions import DistributionFunctions
+from ..Input.distribution.utils import DistributionFunctions
 from ..Input.latticeConfiguration.latticeMain import parameter_input_checker_for_lattice
 
 server, state, ctrl = setup_server()
